Fix loop-rate selftest overflow: 32-bit *1000 wrapped on any real gap
Main.cpp's 't' selftest command computes loop() throughput as
(loop_counter delta) * 1000 / (ms delta), entirely in uint32_t. This
loop() has been measured at ~320,000 Hz on the bench -- at that rate,
any two 't' presses more than ~13 seconds apart accumulate enough
loop_counter delta that the *1000 multiplication overflows uint32_t
and silently wraps, printing a plausible-looking but wrong number.
Since nobody presses 't' twice within 3 seconds during real bench
work, this line has likely been showing garbage in most realistic use
since it was written.

Confirmed on hardware 2026-09-03: a ~3s gap read 320286 Hz; a ~55s gap
(after the fix) read 322198 Hz -- consistent within 0.6%. Before the
fix, gaps in the tens of seconds to tens of minutes produced numbers
from 226 Hz to 357928 Hz with no relationship to actual load, matching
predicted overflow-wrap values computed by hand from the surrounding
readings.

Fixed by computing the multiplication in uint64_t before dividing.

+++ software/src/Main.cpp
@@ -778,7 +778,15 @@ FLASHMEM __attribute__((noinline)) static void SelfTest() {
     static uint32_t last_lc = 0, last_ms = 0;
     const uint32_t lc = loop_counter, ms = millis();
     if (last_ms && ms != last_ms)
-      Serial.printf("loop rate ~%lu Hz\n", (lc - last_lc) * 1000 / (ms - last_ms));
+      // loop() has been observed north of 300 kHz between UI/audio work; at
+      // that rate any gap over ~12-40s between two 't' presses overflowed
+      // this in 32-bit arithmetic ((lc-last_lc)*1000 wrapping uint32_t) and
+      // printed a silently-wrong number -- confirmed on the bench 2026-09-03
+      // (a 92s gap read ~29802 Hz where the true rate, from an adjacent
+      // ~3s-gap reading, was ~357928 Hz; the wrapped math matches exactly).
+      // 64-bit intermediate math avoids it regardless of interval length.
+      Serial.printf("loop rate ~%lu Hz\n",
+                    (unsigned long)((uint64_t)(lc - last_lc) * 1000 / (ms - last_ms)));
     last_lc = lc; last_ms = ms;
     const uint32_t t0 = OC::CORE::ticks;
     delay(5);
